feat: remove dashboard tab bar and split into dedicated route pages
- Remove tab bar UI from dashboard page
- Create /settings page with profile/identity/password forms
- Create /received-bookings page with trips and reservations management
- Dashboard (/dashboard) now shows only sent bookings overview
- Update sidebar links to point to new dedicated routes
- All business logic preserved in respective page components
