+++ front-office/src/app/pages/dashboard/dashboard-page.component.ts
@@ -1,106 +1,22 @@
-import { ChangeDetectorRef, Component, ElementRef, ViewChild, inject, OnInit } from '@angular/core';
+import { Component, inject, OnInit } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { FormsModule, ReactiveFormsModule, FormBuilder, Validators } from '@angular/forms';
-import { ActivatedRoute, Router, RouterLink } from '@angular/router';
-import { forkJoin, of } from 'rxjs';
-import { catchError } from 'rxjs/operators';
-import * as htmlToImage from 'html-to-image';
+import { Router } from '@angular/router';
 import { AuthService } from '../../services/auth.service';
-import { ShareCardData } from '../../models/share-card.model';
 import { TripService } from '../../services/trip.service';
-import { Trip } from '../../models/trip.model';
 import { BookingResponse } from '../../models/booking.model';
-import { UpdateProfileRequest } from '../../models/auth.model';
-import { ShareCardMapperService } from '../../services/share-card-mapper.service';
-import { ShareCardStoryComponent } from '../../components/share-card-story/share-card-story.component';
-import { TravelerTripsDesktopListComponent } from '../../components/dashboard/traveler-trips-desktop-list/traveler-trips-desktop-list.component';
-import { DashboardReceivedMobileCardComponent } from '../../components/dashboard/dashboard-received-mobile-card/dashboard-received-mobile-card.component';
-
-type Tab = 'profile' | 'chats' | 'received' | 'sent';
 
 @Component({
   selector: 'app-dashboard-page',
   standalone: true,
   imports: [
     CommonModule,
-    FormsModule,
-    ReactiveFormsModule,
-    RouterLink,
-    ShareCardStoryComponent,
-    TravelerTripsDesktopListComponent,
-    DashboardReceivedMobileCardComponent,
   ],
   templateUrl: './dashboard-page.component.html',
 })
 export class DashboardPageComponent implements OnInit {
   private readonly authService = inject(AuthService);
-  private readonly cdr = inject(ChangeDetectorRef);
-  private readonly shareCardMapperService = inject(ShareCardMapperService);
   private readonly tripService = inject(TripService);
   private readonly router = inject(Router);
-  private readonly route = inject(ActivatedRoute);
-  private readonly fb = inject(FormBuilder);
-  @ViewChild('shareCardCapture') private shareCardCapture?: ElementRef<HTMLElement>;
-
-  activeTab: Tab = 'profile';
-
-  /** Tab definitions with labels and inline SVG icons. */
-  tabs = [
-    { id: 'profile' as Tab, label: 'Mon profil', icon: '<svg class="w-4 h-4" fill="none" stroke="currentColor" stroke-width="2" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" d="M16 7a4 4 0 11-8 0 4 4 0 018 0zM12 14a7 7 0 00-7 7h14a7 7 0 00-7-7z"/></svg>' },
-    { id: 'chats' as Tab, label: 'Mes chats', icon: '<svg class="w-4 h-4" fill="none" stroke="currentColor" stroke-width="2" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" d="M8 12h.01M12 12h.01M16 12h.01M21 12c0 4.418-4.03 8-9 8a9.863 9.863 0 01-4.255-.949L3 20l1.395-3.72C3.512 15.042 3 13.574 3 12c0-4.418 4.03-8 9-8s9 3.582 9 8z"/></svg>' },
-    { id: 'received' as Tab, label: 'Réservations reçues', icon: '<svg class="w-4 h-4" fill="none" stroke="currentColor" stroke-width="2" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" d="M20 13V6a2 2 0 00-2-2H6a2 2 0 00-2 2v7m16 0v5a2 2 0 01-2 2H6a2 2 0 01-2-2v-5m16 0h-2.586a1 1 0 00-.707.293l-2.414 2.414a1 1 0 01-.707.293h-3.172a1 1 0 01-.707-.293l-2.414-2.414A1 1 0 006.586 13H4"/></svg>' },
-    { id: 'sent' as Tab, label: 'Mes demandes', icon: '<svg class="w-4 h-4" fill="none" stroke="currentColor" stroke-width="2" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" d="M9 5H7a2 2 0 00-2 2v12a2 2 0 002 2h10a2 2 0 002-2V7a2 2 0 00-2-2h-2M9 5a2 2 0 002 2h2a2 2 0 002-2M9 5a2 2 0 012-2h2a2 2 0 012 2"/></svg>' },
-  ];
-
-  // ── Section 1: Informations générales ─────────────────────────────────────
-  infoForm = this.fb.group({
-    firstName: ['', Validators.required],
-    lastName: ['', Validators.required],
-    phone: [''],
-    identityDocument: ['', Validators.required],
-  });
-  infoSuccess = '';
-  infoError = '';
-  isSavingInfo = false;
-
-  // Photo
-  photoPreview: string | null = null;
-  selectedPhotoFile: File | null = null;
-  isUploadingPhoto = false;
-  photoError = '';
-  selectedIdentityProofFile: File | null = null;
-
-  // ── Section 2: Changer l'e-mail ───────────────────────────────────────────
-  emailForm = this.fb.group({
-    newEmail: ['', [Validators.required, Validators.email]],
-  });
-  emailSuccess = '';
-  emailError = '';
-  isSendingEmail = false;
-
-  // ── Section 3: Changer le mot de passe ────────────────────────────────────
-  passwordForm = this.fb.group({
-    oldPassword: ['', Validators.required],
-    newPassword: ['', [Validators.required, Validators.minLength(8)]],
-    confirmPassword: ['', Validators.required],
-  });
-  passwordSuccess = '';
-  passwordError = '';
-  isSavingPassword = false;
-
-  // ── Received reservations ─────────────────────────────────────────────────
-  myTrips: Trip[] = [];
-  tripBookingsMap: Record<number, BookingResponse[]> = {};
-  isLoadingTrips = false;
-  receivedTripsActionError = '';
-  isCancellingTrip = false;
-  isDeletingTrip = false;
-  isCompletingTrip = false;
-  completingTripId: number | null = null;
-  isGeneratingShareCard = false;
-  generatingShareCardTripId: number | null = null;
-  shareCardError = '';
-  shareCardData: ShareCardData | null = null;
 
   // ── Sent requests ─────────────────────────────────────────────────────────
   myBookings: BookingResponse[] = [];
@@ -114,321 +30,7 @@ export class DashboardPageComponent implements OnInit {
       this.router.navigate(['/login']);
       return;
     }
-    const user = this.authService.getUser();
-    if (user) {
-      this.infoForm.patchValue({
-        firstName: user.firstName,
-        lastName: user.lastName,
-        phone: user.phone ?? '',
-        identityDocument: user.identityDocument ?? '',
-      });
-      this.emailForm.patchValue({
-        newEmail: user.email,
-      });
-      this.photoPreview = user.photoUrl ?? null;
-    }
-
-    this.route.queryParamMap.subscribe((params) => {
-      const tab = params.get('tab');
-      if (this.isTab(tab)) this.setTab(tab);
-    });
-  }
-
-  needsIdentityDocument(): boolean {
-    return !this.authService.getUser()?.identityDocument?.trim();
-  }
-
-  hasLocalPassword(): boolean {
-    return this.authService.getUser()?.hasPassword !== false;
-  }
-
-  setTab(tab: Tab): void {
-    this.activeTab = tab;
-    if (tab === 'received' && this.myTrips.length === 0) this.loadMyTrips();
-    if (tab === 'sent' && this.myBookings.length === 0) this.loadSentBookings();
-  }
-
-  private isTab(value: string | null): value is Tab {
-    return value === 'profile' || value === 'chats' || value === 'received' || value === 'sent';
-  }
-
-  // ── Photo ─────────────────────────────────────────────────────────────────
-  onPhotoSelected(event: Event): void {
-    const input = event.target as HTMLInputElement;
-    if (!input.files?.length) return;
-    const file = input.files[0];
-    this.selectedPhotoFile = file;
-    const reader = new FileReader();
-    reader.onload = (e) => { this.photoPreview = e.target?.result as string; };
-    reader.readAsDataURL(file);
-    this.uploadPhoto();
-  }
-
-  uploadPhoto(): void {
-    const user = this.authService.getUser();
-    if (!user || !this.selectedPhotoFile) return;
-    this.isUploadingPhoto = true;
-    this.photoError = '';
-    this.authService.uploadPhoto(user.id, this.selectedPhotoFile).subscribe({
-      next: () => { this.isUploadingPhoto = false; },
-      error: () => { this.photoError = "Erreur lors de l'upload de la photo."; this.isUploadingPhoto = false; },
-    });
-  }
-
-  onIdentityProofSelected(event: Event): void {
-    const input = event.target as HTMLInputElement;
-    this.selectedIdentityProofFile = input.files?.length ? input.files[0] : null;
-  }
-
-  // ── Infos générales ──────────────────────────────────────────────────────
-  saveInfo(): void {
-    if (this.infoForm.invalid || this.isSavingInfo) return;
-    const user = this.authService.getUser();
-    if (!user) return;
-    this.isSavingInfo = true;
-    this.infoSuccess = '';
-    this.infoError = '';
-    const val = this.infoForm.value;
-    const payload: UpdateProfileRequest = {
-      firstName: val.firstName ?? undefined,
-      lastName: val.lastName ?? undefined,
-      phone: val.phone || undefined,
-      identityDocument: val.identityDocument || undefined,
-    };
-    this.authService.updateProfile(user.id, payload).subscribe({
-      next: () => { this.infoSuccess = 'Informations mises à jour.'; this.isSavingInfo = false; },
-      error: () => { this.infoError = 'Erreur lors de la mise à jour.'; this.isSavingInfo = false; },
-    });
-  }
-
-  // ── Email ────────────────────────────────────────────────────────────────
-  sendEmailChange(): void {
-    if (this.emailForm.invalid || this.isSendingEmail) return;
-    const user = this.authService.getUser();
-    if (!user) return;
-    this.isSendingEmail = true;
-    this.emailSuccess = '';
-    this.emailError = '';
-    const newEmail = this.emailForm.value.newEmail!;
-    this.authService.requestEmailChange(user.id, newEmail).subscribe({
-      next: () => {
-        this.emailSuccess = `Un lien de confirmation a été envoyé à ${newEmail}. Vérifiez votre boîte mail.`;
-        this.isSendingEmail = false;
-      },
-      error: () => { this.emailError = "Erreur lors de l'envoi. L'adresse est peut-être déjà utilisée."; this.isSendingEmail = false; },
-    });
-  }
-
-  // ── Password ─────────────────────────────────────────────────────────────
-  savePassword(): void {
-    if (this.passwordForm.invalid || this.isSavingPassword) return;
-    const val = this.passwordForm.value;
-    if (val.newPassword !== val.confirmPassword) {
-      this.passwordError = 'Les nouveaux mots de passe ne correspondent pas.';
-      return;
-    }
-    const user = this.authService.getUser();
-    if (!user) return;
-    this.isSavingPassword = true;
-    this.passwordSuccess = '';
-    this.passwordError = '';
-    this.authService.changePassword(user.id, val.oldPassword!, val.newPassword!).subscribe({
-      next: () => {
-        this.passwordSuccess = 'Mot de passe mis à jour avec succès.';
-        this.passwordForm.reset();
-        this.isSavingPassword = false;
-      },
-      error: () => { this.passwordError = 'Ancien mot de passe incorrect ou erreur serveur.'; this.isSavingPassword = false; },
-    });
-  }
-
-  saveProfileChanges(): void {
-    const user = this.authService.getUser();
-    if (!user) return;
-
-    this.saveInfo();
-
-    const nextEmail = this.emailForm.value.newEmail?.trim();
-    if (nextEmail && nextEmail !== user.email) {
-      this.sendEmailChange();
-    }
-
-    if (this.hasLocalPassword()) {
-      const val = this.passwordForm.value;
-      const hasPasswordInput = !!(val.oldPassword || val.newPassword || val.confirmPassword);
-      if (hasPasswordInput) {
-        this.savePassword();
-      }
-    }
-  }
-
-  resetProfileForms(): void {
-    const user = this.authService.getUser();
-    if (!user) return;
-
-    this.infoForm.reset({
-      firstName: user.firstName,
-      lastName: user.lastName,
-      phone: user.phone ?? '',
-      identityDocument: user.identityDocument ?? '',
-    });
-    this.emailForm.reset({
-      newEmail: user.email,
-    });
-    this.passwordForm.reset();
-    this.selectedIdentityProofFile = null;
-    this.infoSuccess = '';
-    this.infoError = '';
-    this.emailSuccess = '';
-    this.emailError = '';
-    this.passwordSuccess = '';
-    this.passwordError = '';
-  }
-
-  // ── Received reservations ────────────────────────────────────────────────
-  loadMyTrips(): void {
-    this.isLoadingTrips = true;
-    this.receivedTripsActionError = '';
-    this.tripService.getMyTrips().subscribe({
-      next: (trips) => {
-        this.myTrips = trips;
-        if (trips.length === 0) { this.isLoadingTrips = false; return; }
-        const requests = trips.map((t) => this.tripService.getTripBookings(t.id).pipe(catchError(() => of([]))));
-        forkJoin(requests).subscribe((results) => {
-          trips.forEach((t, i) => { this.tripBookingsMap[t.id] = results[i] as BookingResponse[]; });
-          this.isLoadingTrips = false;
-        });
-      },
-      error: () => { this.isLoadingTrips = false; },
-    });
-  }
-
-  selectTrip(tripId: number): void {
-    void this.router.navigate(['/trips', tripId, 'reservations']);
-  }
-
-  async downloadShareCardPng(tripId?: number): Promise<void> {
-    const trip = typeof tripId === 'number'
-      ? this.myTrips.find((currentTrip) => currentTrip.id === tripId)
-      : undefined;
-    const user = this.authService.getUser();
-
-    if (!trip || trip.status !== 'ACTIVE' || !user) {
-      this.shareCardError = 'Sélectionnez un voyage actif pour générer la carte.';
-      return;
-    }
-
-    this.shareCardError = '';
-    this.shareCardData = this.shareCardMapperService.mapActiveTripToShareCard(trip, user);
-    this.cdr.detectChanges();
-    this.isGeneratingShareCard = true;
-    this.generatingShareCardTripId = trip.id;
-
-    try {
-      await new Promise<void>((resolve) => {
-        requestAnimationFrame(() => resolve());
-      });
-
-      const captureElement = this.shareCardCapture?.nativeElement;
-      if (!captureElement) {
-        throw new Error('Share card capture element missing');
-      }
-
-      const pngDataUrl = await this.generatePngFromElement(captureElement);
-
-      const downloadLink = document.createElement('a');
-      downloadLink.href = pngDataUrl;
-      downloadLink.download = `colick-carte-partage-${this.shareCardMapperService.buildFileDate(trip.departureTime)}.png`;
-      downloadLink.click();
-    } catch {
-      this.shareCardError = 'Impossible de générer la carte PNG pour le moment.';
-    } finally {
-      this.isGeneratingShareCard = false;
-      this.generatingShareCardTripId = null;
-    }
-  }
-
-  editTrip(tripId: number): void {
-    void this.router.navigate(['/propose', tripId]);
-  }
-
-  completeTrip(tripId: number): void {
-    if (this.isCompletingTrip) {
-      return;
-    }
-
-    this.isCompletingTrip = true;
-    this.completingTripId = tripId;
-    this.receivedTripsActionError = '';
-    this.tripService.completeTrip(tripId).subscribe({
-      next: (updatedTrip) => {
-        const tripIndex = this.myTrips.findIndex((trip) => trip.id === updatedTrip.id);
-        if (tripIndex >= 0) {
-          this.myTrips[tripIndex] = updatedTrip;
-        }
-        this.isCompletingTrip = false;
-        this.completingTripId = null;
-      },
-      error: (err: { error?: { message?: string } }) => {
-        this.receivedTripsActionError = err.error?.message || "Erreur lors du passage du trajet à l'état effectué.";
-        this.isCompletingTrip = false;
-        this.completingTripId = null;
-      },
-    });
-  }
-
-  handleDesktopTripShareCardDownload(tripId: number): void {
-    void this.downloadShareCardPng(tripId);
-  }
-
-  handleDesktopTripCompletion(tripId: number): void {
-    this.completeTrip(tripId);
-  }
-
-  handleDesktopTripEdit(tripId: number): void {
-    this.editTrip(tripId);
-  }
-
-  handleDesktopTripDelete(tripId: number): void {
-    this.deleteTrip(tripId);
-  }
-
-  handleDesktopTripCancel(tripId: number): void {
-    this.cancelTrip(tripId);
-  }
-
-  deleteTrip(tripId: number): void {
-    if ((this.tripBookingsMap[tripId] ?? []).length > 0) {
-      this.receivedTripsActionError = 'Ce trajet ne peut pas être supprimé car il a déjà reçu des demandes.';
-      return;
-    }
-
-    if (!confirm('Êtes-vous sûr de vouloir supprimer ce trajet ?')) return;
-    this.isDeletingTrip = true;
-    this.receivedTripsActionError = '';
-    this.tripService.cancelTrip(tripId).subscribe({
-      next: () => {
-        this.removeTripFromState(tripId);
-        this.isDeletingTrip = false;
-      },
-      error: () => {
-        this.receivedTripsActionError = 'Erreur lors de la suppression du trajet.';
-        this.isDeletingTrip = false;
-      },
-    });
-  }
-
-  cancelTrip(tripId: number): void {
-    if (!confirm('Êtes-vous sûr de vouloir annuler ce trajet ? Tous les demandeurs seront notifiés.')) return;
-    this.isCancellingTrip = true;
-    this.receivedTripsActionError = '';
-    this.tripService.cancelTrip(tripId).subscribe({
-      next: () => {
-        this.removeTripFromState(tripId);
-        this.isCancellingTrip = false;
-      },
-      error: () => { this.receivedTripsActionError = "Erreur lors de l'annulation du trajet."; this.isCancellingTrip = false; },
-    });
+    this.loadSentBookings();
   }
 
   loadSentBookings(): void {
@@ -460,25 +62,5 @@ export class DashboardPageComponent implements OnInit {
 
   validationChannelLabel(channel?: string): string {
     return channel === 'SMS' ? 'SMS' : 'e-mail';
-  }
-
-  tripStatusLabel(status: Trip['status']): string {
-    return ({ ACTIVE: 'Actif', COMPLETED: 'Terminé', CANCELLED: 'Annulé' } as Record<Trip['status'], string>)[status];
-  }
-
-  tripStatusClass(status: Trip['status']): string {
-    return ({ ACTIVE: 'bg-accent/10 text-accent', COMPLETED: 'bg-success/10 text-success', CANCELLED: 'bg-background-primary text-text-muted' } as Record<Trip['status'], string>)[status];
-  }
-
-  private generatePngFromElement(element: HTMLElement): Promise<string> {
-    return htmlToImage.toPng(element, {
-      cacheBust: true,
-      pixelRatio: 2,
-    });
-  }
-
-  private removeTripFromState(tripId: number): void {
-    this.myTrips = this.myTrips.filter((trip) => trip.id !== tripId);
-    delete this.tripBookingsMap[tripId];
   }
 }
